added support for DockerInDocker (DinD)
