+++ docs/doc_example/CMakeLists.txt
@@ -89,7 +89,6 @@ add_onnx_mlir_executable(OMRuntimeCppTest
 
   OMExecutionSession
   OMCompile
-  OMUnifiedCompile
   )
 
 set_output_directory(OMRuntimeCppTest BINARY_DIR ${CMAKE_CURRENT_BINARY_DIR} LIBRARY_DIR ${CMAKE_CURRENT_BINARY_DIR})
