+++ src/Compiler/OMCompile.hpp
@@ -17,6 +17,8 @@
 #ifndef ONNX_MLIR_COMPILER_SESSION
 #define ONNX_MLIR_COMPILER_SESSION
 
+#include <map>
+#include <memory>
 #include <stdexcept>
 #include <string>
 #include <vector>
@@ -26,7 +28,10 @@
 // (except implicitly).
 namespace onnx_mlir {
 
-// Exception class
+// Forward declarations.
+class Command;
+
+// Exception class.
 class OMCompileException : public std::runtime_error {
 public:
   explicit OMCompileException(const std::string &msg)
@@ -35,30 +40,38 @@ public:
 
 /**
  * @class OMCompile
- * @brief C++ interface for compiling ONNX models using the onnx-mlir compiler.
+ * @brief Unified C++ interface for compiling ONNX models locally or in containers.
  *
  * This class provides a thread-safe interface to compile ONNX models from files
  * (.onnx, .mlir, or .onnxtext formats) into various output formats such as
  * shared libraries (.so/.dll), object files (.o/.obj), or JAR files.
  *
+ * ## Compilation Modes
+ * The class supports three compilation modes:
+ * 1. **Local compilation** - Direct execution of onnx-mlir binary
+ * 2. **Container compilation** - onnx-mlir running in Docker/Podman
+ * 3. **Docker-in-Docker (DinD)** - Automatic detection and handling
+ *
  * ## Thread Safety
  * This interface is thread-safe and does not read any flags from environment
- * variables. All compilation options must be explicitly passed via the flags
- * parameter.
+ * variables (except for DinD configuration). All compilation options must be
+ * explicitly passed via the flags parameter.
  *
  * ## Compilation Process
- * The class invokes the onnx-mlir compiler executable to perform the actual
- * compilation. When generating libraries or JAR files, the compiler
- * automatically links in the required lightweight runtime libraries.
+ * The class invokes the onnx-mlir compiler executable (locally or in container)
+ * to perform the actual compilation. When generating libraries or JAR files,
+ * the compiler automatically links in the required lightweight runtime libraries.
  *
  * ## Runtime Library Location
  * By default, runtime libraries are expected in system-wide directories
  * (typically /usr/local/lib). To use a custom location, set the
  * ONNX_MLIR_LIBRARY_PATH environment variable before compilation.
  *
- * ## Usage Example
+ * ## Usage Examples
+ *
+ * ### Local Compilation
  * @code
- *   OMCompile session;
+ *   OMCompile session;  // Default constructor for local mode
  *   try {
  *     session.compile("model.onnx", "-O3 -o output");
  *     std::string outputFile = session.getOutputFilename();
@@ -66,6 +79,27 @@ public:
  *   } catch (const OMCompileException& e) {
  *     std::cerr << "Compilation failed: " << e.what() << std::endl;
  *   }
+ * @endcode
+ *
+ * ### Container Compilation
+ * @code
+ *   // Use containerized onnx-mlir
+ *   OMCompile session(
+ *       "ghcr.io/onnxmlir/onnx-mlir-dev",  // image
+ *       "/workdir/onnx-mlir/build/Debug/bin/onnx-mlir"  // compiler in container
+ *   );
+ *   session.compile("model.onnx", "-O3");
+ * @endcode
+ *
+ * ### Docker-in-Docker (DinD)
+ * @code
+ *   // DinD is automatically detected - no special code needed!
+ *   // Just ensure Docker socket is mounted when running outer container:
+ *   // docker run -v /var/run/docker.sock:/var/run/docker.sock ...
+ *
+ *   OMCompile session("image", "compiler");
+ *   session.compile("model.onnx", "-O3");
+ *   // Paths are automatically resolved for host access
  * @endcode
  *
  * ## Supported Compiler Flags
@@ -83,13 +117,83 @@ public:
  */
 class OMCompile {
 public:
+  /// Container engine type.
+  enum class ContainerEngine { Docker, Podman, Auto };
+
+  /// Compilation mode.
+  enum class CompilationMode { Local, Container };
+
   /**
-   * @brief Default constructor.
+   * @brief Default constructor for local compilation.
    *
-   * Creates a OMCompile object. Actual compilation is deferred until
+   * Creates an OMCompile object for local compilation using the default
+   * onnx-mlir binary from PATH. Actual compilation is deferred until
    * the compile() method is called.
    */
-  OMCompile() = default;
+  OMCompile();
+
+  /**
+   * @brief Constructor for local compilation with custom compiler path.
+   *
+   * @param compilerPath Path to local onnx-mlir binary (empty uses PATH default)
+   * @param verbose Enable verbose output (default: false)
+   *
+   * @code
+   *   // Use onnx-mlir from PATH
+   *   OMCompile compiler({});
+   *
+   *   // Use specific compiler
+   *   OMCompile compiler("/path/to/onnx-mlir");
+   *
+   *   // With verbose
+   *   OMCompile compiler({}, true);
+   * @endcode
+   */
+  explicit OMCompile(const std::string &compilerPath, bool verbose = false);
+
+  /**
+   * @brief Constructor for container-based compilation.
+   *
+   * Performs one-time setup:
+   * - Detects container engine (docker/podman) if Auto
+   * - Verifies/pulls container image
+   * - Auto-detects compiler path for known images
+   * - Automatically detects and handles Docker-in-Docker (DinD) scenarios
+   *
+   * @param containerImage Container image name (required, but {} uses first known image)
+   * @param compilerPathInContainer Path to compiler in container (required, but {} auto-detects)
+   * @param engine Container engine to use (default: Auto - auto-detect)
+   * @param autoPull Automatically pull missing images (default: true)
+   * @param verbose Enable verbose output (default: false)
+   *
+   * @note Docker-in-Docker Support:
+   * When running inside a container, this class automatically detects the DinD
+   * scenario and adjusts volume mount paths accordingly. For DinD to work:
+   * - Mount Docker socket: -v /var/run/docker.sock:/var/run/docker.sock
+   * - Use absolute paths for model files
+   * - Ensure paths are accessible from the host (not just outer container)
+   *
+   * Environment variables for advanced DinD configuration:
+   * - OM_DOCKER_HOST_PATH_PREFIX: Prefix to add to paths (e.g., "/host")
+   * - OM_DIND_DISABLE: Set to "1" to disable DinD detection
+   *
+   * @code
+   *   // Use defaults (first known image, auto-detect compiler, auto-detect engine)
+   *   OMCompile compiler({}, {});
+   *
+   *   // Specific image with auto-detected compiler path
+   *   OMCompile compiler("ghcr.io/onnxmlir/onnx-mlir", {});
+   *
+   *   // With verbose mode (shows DinD detection info)
+   *   OMCompile compiler({}, {}, OMCompile::ContainerEngine::Auto, true, true);
+   *
+   *   // Docker-in-Docker works automatically - no special configuration needed
+   * @endcode
+   */
+  OMCompile(const std::string &containerImage,
+      const std::string &compilerPathInContainer,
+      ContainerEngine engine = ContainerEngine::Auto, bool autoPull = true,
+      bool verbose = false);
 
   /**
    * @brief Destructor.
@@ -99,8 +203,9 @@ public:
   /**
    * @brief Compile an ONNX model with specified flags.
    *
-   * Invokes the onnx-mlir compiler to compile the input model. The method
-   * blocks until compilation completes or fails.
+   * Invokes the onnx-mlir compiler (locally or in container) to compile the
+   * input model. The method blocks until compilation completes or fails.
+   * Works identically for both local and container modes.
    *
    * @param modelPath Path to the input model file (.onnx, .mlir, or .onnxtext).
    * Can include a directory path. If empty, the flags parameter must contain
@@ -108,8 +213,8 @@ public:
    * @param flags Compilation flags as a single string (e.g., "-O3 -o output").
    *              Supports quoted strings for paths with spaces.
    * @param compilerPath Optional path to the compiler binary, including the
-   * binary name.  If empty (default) standard onnx-mlir binary will be used at
-   * standard location.
+   * binary name. If empty (default) standard onnx-mlir binary will be used at
+   * standard location. Only used in local mode.
    * @param logFilename Optional path to a file where compilation logs will be
    *                    written. If empty, logs go to stdout/stderr.
    *
@@ -174,6 +279,32 @@ public:
   bool hasOutputConstantFilename() { return !outputConstantFilename.empty(); }
 
   /**
+   * @brief Get the compilation mode.
+   */
+  CompilationMode getMode() const { return mode; }
+
+  /**
+   * @brief Get the container engine name (only for container mode).
+   */
+  std::string getContainerEngineName() const;
+
+  /**
+   * @brief Check if currently running inside a container (Docker-in-Docker).
+   *
+   * Detects Docker-in-Docker scenarios by checking:
+   * - Presence of /.dockerenv file (Docker-specific marker)
+   * - Container indicators in /proc/1/cgroup (works for Docker and Podman)
+   * - /run/.containerenv file (Podman-specific marker)
+   * - Hostname pattern + cgroups v2 (modern Podman)
+   *
+   * This detection is cached after first call for performance.
+   * Can be disabled by setting OM_DIND_DISABLE=1 environment variable.
+   *
+   * @return true if running in a container, false otherwise
+   */
+  bool isRunningInContainer() const;
+
+  /**
    * @brief Static helper to extract input filename from model path and flags.
    *
    * Useful for determining the input file before compilation, especially when
@@ -211,15 +342,53 @@ public:
   static std::string getModelTag(const std::string &flags);
 
 private:
-  /// Parsed compilation flags as a vector of individual arguments
+  // Compilation mode.
+  const CompilationMode mode;
+
+  // Common configuration.
+  const bool verbose;
+
+  // Local mode configuration.
+  const std::string localCompilerPath;
+
+  // Container mode configuration.
+  std::string containerImage;          // Not const - set in constructor body.
+  std::string compilerPathInContainer; // Not const - set in constructor body.
+  const ContainerEngine containerEngine;
+  const bool autoPullImage;
+  std::string detectedEngineName; // "docker" or "podman".
+
+  // Docker-in-Docker state (cached detection).
+  mutable bool dindDetected = false;      // Cached DinD detection result.
+  mutable bool dindDetectionDone = false; // Whether detection has been performed.
+
+  // Compilation state.
+  /// Parsed compilation flags as a vector of individual arguments.
   std::vector<std::string> flagVect;
 
-  /// Absolute path to the output file from the last successful compilation
+  /// Absolute path to the output file from the last successful compilation.
   std::string outputFilename = {};
   std::string outputConstantFilename = {};
 
-  /// Flag indicating whether the last compilation completed successfully
+  /// Flag indicating whether the last compilation completed successfully.
+  bool successfullyInitialized = false; // Set to true when constructor completes.
   bool successfullyCompiled = false;
+
+  // Known container image configurations.
+  static const std::map<std::string, std::string> knownImageConfigs;
+
+  // Helper methods for container mode.
+  void verifyContainerSetup();
+
+  // Core compilation logic - separate methods for clarity.
+  std::unique_ptr<Command> createLocalCompileCommand(
+      const std::string &modelPath, const std::vector<std::string> &flagVect,
+      const std::string &inputFilename);
+
+  std::unique_ptr<Command> createContainerCompileCommand(
+      const std::vector<std::string> &flagVect,
+      const std::string &inputFilename, const std::string &modelDir,
+      const std::string &outputDir);
 };
 
 } // namespace onnx_mlir
