+++ src/Runtime/python/PyFloat16.hpp
@@ -66,4 +66,3 @@ private:
 
 #endif // ONNX_MLIR_PY_FLOAT16_H
 
-// Made with Bob
